# Improve the performance of `get_predict()`