--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: marginaleffects
 Title: Predictions, Comparisons, Slopes, Marginal Means, and Hypothesis Tests
-Version: 0.9.0
+Version: 0.9.0.9000
 Authors@R: 
     c(person(given = "Vincent",
              family = "Arel-Bundock",
@@ -16,7 +16,11 @@ Authors@R:
              family = "Greifer",
              role = "ctb",
              email = "noah.greifer@gmail.com",
-             comment = c(ORCID = "0000-0003-3067-7154")))
+             comment = c(ORCID = "0000-0003-3067-7154")),
+      person("Etienne", "Bacher",
+             email = "etienne.bacher@protonmail.com",
+             role = "ctb",
+             comment = c(ORCID = "0000-0002-9271-5075"))))
 Description: Compute and plot predictions, slopes, marginal means, and comparisons (contrasts, risk ratios, odds, etc.) for over 70 classes of statistical models in R. Conduct linear and non-linear hypothesis tests, or equivalence tests. Calculate uncertainty estimates using the delta method, bootstrapping, or simulation-based inference.
 License: GPL (>= 3)
 Copyright: inst/COPYRIGHTS