Two things:

1. Creating a `data.table()` takes some time and is not necessary in `get_predict()` because there’s no `data.table`-specific operations. It is faster to create a list, which is also accepted by `setDF()`. 
2. If I'm not mistaken, the only reason why you use `as.numeric(pred)` is because of `svyglm` which returns an object of class `svystat`. It's faster to directly assign the class "numeric" when there are a lot of observations.

These improve the performance mostly when `vcov = TRUE`.

I use the same setup as in the [performance vignette](https://vincentarelbundock.github.io/marginaleffects/articles/performance.html) but with less observations because it takes quite some time to run all these benchmarks.

### Before

``` r
library(marginaleffects)

# simulate data and fit a large model
N <- 3*1e4
dat <- data.frame(matrix(rnorm(N * 26), ncol = 26))
mod <- lm(X1 ~ ., dat)

results <- bench::mark(
  # marginal effects at the mean; no standard error
  slopes(mod, vcov = FALSE, newdata = "mean"),
  # marginal effects at the mean
  slopes(mod, newdata = datagrid()),
  # 1 variable; no standard error
  slopes(mod, vcov = FALSE, variables = "X3"),
  # 1 variable
  slopes(mod, variables = "X3"),
  # 26 variables; no standard error
  slopes(mod, vcov = FALSE),
  # 26 variables
  slopes(mod),
  iterations = 10,
  check = FALSE
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.

results[, c(1, 3, 5)]
#> # A tibble: 6 × 3
#>   expression                                    median mem_alloc
#>   <bch:expr>                                  <bch:tm> <bch:byt>
#> 1 slopes(mod, vcov = FALSE, newdata = "mean") 419.16ms  415.82MB
#> 2 slopes(mod, newdata = datagrid())           668.29ms  421.67MB
#> 3 slopes(mod, vcov = FALSE, variables = "X3") 305.74ms  212.78MB
#> 4 slopes(mod, variables = "X3")                  2.06s    1.23GB
#> 5 slopes(mod, vcov = FALSE)                      6.71s    4.13GB
#> 6 slopes(mod)                                   41.76s   28.59GB
```

### After

``` r
results[, c(1, 3, 5)]
#> # A tibble: 6 × 3
#>   expression                                    median mem_alloc
#>   <bch:expr>                                  <bch:tm> <bch:byt>
#> 1 slopes(mod, vcov = FALSE, newdata = "mean") 350.68ms  417.65MB
#> 2 slopes(mod, newdata = datagrid())           611.66ms  420.85MB
#> 3 slopes(mod, vcov = FALSE, variables = "X3") 210.24ms  211.24MB
#> 4 slopes(mod, variables = "X3")                  1.44s     1.2GB
#> 5 slopes(mod, vcov = FALSE)                      5.38s    4.09GB
#> 6 slopes(mod)                                   28.99s    27.8GB
```