@vincentarelbundock 2023-02-02
BTW, ignore the `pkgdown` failure. I think we're exceeding the allotted Github computing resources (not sure; need to investigate).